Unify webui_utils.h and brave_webui_source.h
