+++ browser/ui/BUILD.gn
@@ -42,8 +42,6 @@ source_set("ui") {
     "webui/brave_webui_source.h",
     "webui/webcompat_reporter_ui.cc",
     "webui/webcompat_reporter_ui.h",
-    "webui/webui_utils.cc",
-    "webui/webui_utils.h",
   ]
 
   if (ipfs_enabled) {
