+++ browser/ui/webui/new_tab_page/brave_new_tab_ui.cc
@@ -8,9 +8,9 @@
 #include <string>
 
 #include "brave/browser/new_tab/new_tab_shows_options.h"
+#include "brave/browser/ui/webui/brave_webui_source.h"
 #include "brave/browser/ui/webui/new_tab_page/brave_new_tab_message_handler.h"
 #include "brave/browser/ui/webui/new_tab_page/instant_service_message_handler.h"
-#include "brave/browser/ui/webui/webui_utils.h"
 #include "brave/components/brave_new_tab/resources/grit/brave_new_tab_generated_map.h"
 #include "chrome/browser/profiles/profile.h"
 #include "components/grit/brave_components_resources.h"
